bootstrap-files: use old files from f85b2a1
This is not the bootstrap files from glibc 2.28-era because nixpkgs
changed too much since then. This is from the glibc 2.31-era. After the
bootstrap process, no dependency from these bootstrap files remains, so
we could even use today's bootstrap files if not for this issue when
building older glibc with make >=4.4:
https://sourceware.org/pipermail/libc-help/2023-September/006486.html